Data.Monoid.Instances.Concat

Description

This module defines the monoid transformer data type Concat.

Documentation

data Concat a

Concat is a transparent monoid transformer. The behaviour of the Concat a instances of monoid subclasses is identical to the behaviour of their a instances, up to the pure isomorphism.

The only purpose of Concat then is to change the performance characteristics of various operations. Most importantly, injecting a monoid into Concat has the effect of making mappend a constant-time operation. The splitPrimePrefix and splitPrimeSuffix operations are amortized to constant time, provided that only one or the other is used. Using both operations alternately will trigger the worst-case behaviour of O(n).

concatenate :: PositiveMonoid a => Seq a -> Concat a

Deprecated: Concat is not wrapping Seq any more, don't use concatenate nor extract.

extract :: Concat a -> Seq a

Deprecated: Concat is not wrapping Seq any more, don't use concatenate nor extract.

force :: Monoid a => Concat a -> a